+++ example/src/main/java/org/pfruender/mockinjector/example/ResultOfDependentServiceOne.java
@@ -13,8 +13,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.example;
+package org.pfruender.mockinjector.example;
 
-public class ResultOfDependentServiceTwo {
+public class ResultOfDependentServiceOne {
 
 }
