+++ example/src/main/java/org/pfruender/mockinjector/example/ResultOfDependentServiceTwo.java
@@ -13,9 +13,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.example;
+package org.pfruender.mockinjector.example;
 
-public interface SampleDependentServiceOne {
+public class ResultOfDependentServiceTwo {
 
-  ResultOfDependentServiceOne getResult(ServiceInputParameter param1, ServiceInputParameter param2);
 }
