+++ example/src/main/java/org/pfruender/mockinjector/example/SampleOrchestratingService.java
@@ -13,21 +13,20 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.example;
+package org.pfruender.mockinjector.example;
 
 import javax.inject.Inject;
-import javax.inject.Provider;
 
-public class SampleOrchestratingServiceWithProviders {
-
-  @Inject
-  Provider<SampleDependentServiceOne> sampleDependentServiceOneProvider;
+public class SampleOrchestratingService {
 
   @Inject
-  Provider<SampleDependentServiceTwo> sampleDependentServiceTwoProvider;
+  SampleDependentServiceOne dependentServiceOne;
+
+  @Inject
+  SampleDependentServiceTwo dependentServiceTwo;
 
   public ResultOfDependentServiceTwo doService(ServiceInputParameter param1, ServiceInputParameter param2) {
-    ResultOfDependentServiceOne result1 = sampleDependentServiceOneProvider.get().getResult(param1, param2);
-    return sampleDependentServiceTwoProvider.get().getResult(result1);
+    ResultOfDependentServiceOne result1 = dependentServiceOne.getResult(param1, param2);
+    return dependentServiceTwo.getResult(result1);
   }
 }
