+++ example/src/test/java/org/pfruender/mockinjector/example/SampleOrchestratingServiceUnitTest.java
@@ -13,13 +13,13 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.example;
+package org.pfruender.mockinjector.example;
 
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.pfruender.mockito.MockInjector.injectMocks;
+import static org.pfruender.mockinjector.MockInjector.injectMocks;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.BDDMockito.given;
