+++ src/main/java/org/pfruender/mockinjector/MockInjectorConfigurator.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito;
+package org.pfruender.mockinjector;
 
 import java.lang.annotation.Annotation;
 import java.util.Collection;
