+++ src/test/java/org/pfruender/mockinjector/injection/MyClass.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.injection;
+package org.pfruender.mockinjector.injection;
 
 import javax.annotation.Resource;
 import javax.inject.Inject;
