+++ src/test/java/org/pfruender/mockinjector/provider/IProvided.java
@@ -13,9 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.pfruender.mockito.example;
+package org.pfruender.mockinjector.provider;
 
-public interface SampleDependentServiceTwo {
+public interface IProvided {
 
-  ResultOfDependentServiceTwo getResult(ResultOfDependentServiceOne result1);
+  String getSomething();
+
+  Object getObject();
 }
